+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 * [#102]: Add `icons` property to allow specification of icon components
+* [#103]: Add `title` node property and `showNodeTitle` tree property
 
 ## [v1.2.4](https://github.com/jakezatecky/react-checkbox-tree/compare/v1.2.3...v1.2.4) (2018-08-29)